+++ src/lib.rs
@@ -61,9 +61,47 @@
 //!     fn main() -> Result<()> {
 //!         let html = "<div id=cat />";
 //!         let json = Dom::parse(html)?.to_json_pretty()?;
-//!         println!("{}", json);
+//!         println!("{}", Dom::parse(html)?);
 //!         Ok(())
 //!     }
+//! ```
+//!
+//! Create a dom manually and print it
+//! ```rust
+//! use lithtml::{Dom, Node, Result};
+//!
+//! fn main() -> Result<()> {
+//!     let mut dom = Dom::new();
+//!     dom.children.push(Node::Comment("Welcome to the test"));
+//!     dom.children.push(Node::parse_json(
+//!         r#"{
+//!           "name": "div",
+//!           "variant": "normal",
+//!           "children": [
+//!             {
+//!               "name": "h1",
+//!               "variant": "normal",
+//!               "children": [
+//!                 "Tjena världen!"
+//!               ]
+//!             },
+//!             {
+//!               "name": "p",
+//!               "variant": "normal",
+//!               "children": [
+//!                 "Tänkte bara informera om att Sverige är bättre än Finland i ishockey."
+//!               ]
+//!             }
+//!           ]
+//!         }"#
+//!     )?);
+//!     dom.children.append(&mut Node::parse(
+//!         r#"<div>Testing</div><p>Multiple elements from node</p>"#,
+//!     )?);
+//!
+//!     println!("{}", dom);
+//!     Ok(())
+//! }
 //! ```
 
 #![allow(clippy::needless_doctest_main)]