+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 
 ## 0.7.0-dev
+- Added possibility to create node and dom manually
+- Renamed errors to warnings within the dom
+- Added formatted html output possibility
 - Removed id as own element variable, it's part of attributes now
 - Removed String allocations for better performance
-- Added formatted html output
 
 ## 0.6.0
 - Added github ci/cd workflows